+++ test/peek_app_sdk/ui/odyssey_test.exs
@@ -167,7 +167,7 @@ defmodule PeekAppSDK.UI.OdysseyTest do
       assert html =~ "<fieldset"
       assert html =~ "class=\"fieldset mb-2\""
       assert html =~ "<label>"
-      assert html =~ "<span class=\"label mb-1\">Time Unit</span>"
+      assert html =~ ~r/<span class="label mb-1">.*Time Unit.*<\/span>/s
       assert html =~ "class=\"inline-flex rounded-lg\""
     end
 
@@ -188,7 +188,7 @@ defmodule PeekAppSDK.UI.OdysseyTest do
       assert html =~ "<fieldset"
       assert html =~ "class=\"fieldset mb-2\""
       assert html =~ "<label>"
-      assert html =~ "<span class=\"label mb-1\">Communication Channel</span>"
+      assert html =~ ~r/<span class="label mb-1">.*Communication Channel.*<\/span>/s
       assert html =~ "Email"
       assert html =~ "Text Message"
       # Should have hidden input for form integration